requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.6.0
python-dotenv==1.0.0
lxml==4.9.3
html2text==2020.1.16
